pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bagley"
version = "7.0.0"
description = "Bagley v7 - Ultimate All-in-One Local AI System"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Proprietary"}
authors = [
    {name = "Bagley Engineering Team"}
]
keywords = ["ai", "llm", "image-generation", "video-generation", "tts", "assistant"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core ML
    "torch>=2.4.0",
    "torchvision>=0.19.0",
    "torchaudio>=2.4.0",
    "transformers>=4.46.0",
    "accelerate>=1.0.0",
    "safetensors>=0.4.0",
    "einops>=0.8.0",
    "flash-attn>=2.6.0",
    "triton>=3.0.0",
    
    # Quantization & Optimization
    "bitsandbytes>=0.44.0",
    "auto-gptq>=0.8.0",
    "optimum>=1.23.0",
    "vllm>=0.6.0",
    
    # Image/Video Processing
    "diffusers>=0.31.0",
    "opencv-python>=4.10.0",
    "pillow>=10.4.0",
    "imageio>=2.35.0",
    "imageio-ffmpeg>=0.5.0",
    
    # Audio/TTS
    "soundfile>=0.12.0",
    "librosa>=0.10.0",
    "sounddevice>=0.5.0",
    "pyaudio>=0.2.14",
    
    # API & Server
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "websockets>=13.0",
    "httpx>=0.27.0",
    "aiofiles>=24.0.0",
    
    # Data Processing
    "datasets>=3.0.0",
    "tokenizers>=0.20.0",
    "sentencepiece>=0.2.0",
    "tiktoken>=0.7.0",
    
    # Utils
    "pydantic>=2.9.0",
    "rich>=13.9.0",
    "typer>=0.12.0",
    "tqdm>=4.66.0",
    "numpy>=1.26.0",
    "scipy>=1.14.0",
    "pyyaml>=6.0.0",
    "omegaconf>=2.3.0",
    
    # PC Integration
    "pyautogui>=0.9.54",
    "psutil>=6.0.0",
    "watchdog>=5.0.0",
    "playwright>=1.48.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "black>=24.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "pre-commit>=4.0.0",
]
training = [
    "deepspeed>=0.15.0",
    "wandb>=0.18.0",
    "tensorboard>=2.18.0",
    "fairscale>=0.4.0",
]

[project.scripts]
bagley = "bagley.cli:main"
bagley-train = "bagley.training.cli:main"
bagley-server = "bagley.api.server:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["bagley*"]

[tool.black]
line-length = 100
target-version = ['py311', 'py312']

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
